Added another Baud Rate, fixed bug
Added 230k as a baud rate option, fixed issue with baud rate data type.

+++ ESPixelStick.h
@@ -82,7 +82,7 @@ typedef struct {
 	
 	/* Renard config  */
 	uint16_t channel_count;		/* Number of channels */
-	uint16_t baud;				/* Baudrate of Serial Port */
+	uint32_t baud;				/* Baudrate of Serial Port */
 } __attribute__((packed)) config_t;
 
 /* Globals */
